+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.3-alpha] - 2022-03-20
+
 - Added RotatingFileHandler and StreamHandler to log both to a file and stdout
 - Added new function to tv show probe to parse datetime with support for known imdb formats
+- Added cache for imdb finder (fetch_show)
 
 ## [0.0.2-alpha] - 2022-03-19